finished enemy generation and basic battle behavior. upon player death, player now returns to spawn

+++ src/main/java/com/skyteeee/tungeon/entities/Place.java
@@ -68,7 +68,7 @@ public class Place extends EntityClass {
     }
 
     public void removeEnemy(Enemy enemy) {
-        enemies.remove(enemy.getId());
+        enemies.remove((Integer) enemy.getId());
     }
 
     @Override
